clean up repo and utilize setup.py files for installing modules more easily.
improve download script to automate it entirely without requiring download directories and make consistent where the data will be downloaded.

add .gitignore to not upload the data nor cityscapeScripts repo.
