+++ .gitignore
@@ -1,0 +1,4 @@
+# ignore all files in data
+data/
+# ignore all files in third-party
+third-party/
