+++ data_prep/setup.py
@@ -1,0 +1,9 @@
+#!/usr/bin/env python
+from distutils.core import setup
+
+setup(
+    name="data_prep",
+    version="0.0",
+    description="Data preparation",
+    packages=["data_prep"],
+)
