+++ utils/setup.py
@@ -1,0 +1,9 @@
+#!/usr/bin/env python
+from distutils.core import setup
+
+setup(
+    name="utils",
+    version="0.0",
+    description="miscellaneous utilities for Project",
+    packages=["utils"]
+)
